json

+++ controllers/seller/product.js
@@ -1,7 +1,7 @@
 'use strict'
-const { Product } = require('../../models')
+const { Product, ProductCategory, Category } = require('../../models')
 const { Op } = require("sequelize");
-const admin = require('../../helpers/firebase')
+const admin = require('../../helpers/firebase');
 const app = require('express')()
 app.locals.bucket = admin.storage().bucket()
 
@@ -9,13 +9,15 @@ class ProductController {
     static async addProduct(req, res, next) {
         try {
             const user_id = req.userData.id
-            const { name, base_price, category, location } = req.body
-            
-
-            var image_url 
+            const { name, description, base_price, category_ids, location } = req.body
+            let categories = category_ids
+            if (typeof categories == 'string') {
+                categories = categories.split(',')
+            }
+            var image_url
             var image_name
             if (!req.files || Object.keys(req.files).length === 0) {
-               image_url = null
+                image_url = null
             } else {
                 let image = req.files.image;
                 image_name = `PR-${Number(new Date())}-${image.name}`
@@ -28,14 +30,26 @@ class ProductController {
                     user_id
                 }
             })
+
             if (products.length < 5) {
-                const newProduct = await Product.create({ name, base_price, category, location, user_id, image_url, image_name })
+                const newProduct = await Product.create({ name, description, base_price, location, user_id, image_url, image_name })
+                Promise.all(
+                    categories.map(async el => {
+
+                        let result = await ProductCategory.create({
+                            category_id: el,
+                            product_id: newProduct.id
+                        })
+                        return result
+                    })
+                )
+
                 res.status(201).json(newProduct)
             } else {
-                next({name: 'maxProducts'})
+                next({ name: 'maxProducts' })
             }
         } catch (err) {
-           next(err)
+            next(err)
         }
     }
     static async getMyProducts(req, res, next) {
@@ -45,7 +59,16 @@ class ProductController {
             const products = await Product.findAll({
                 where: {
                     user_id
-                }
+                },
+                include: [{
+                    model: Category,
+                    through: {
+                        attributes: []
+                    },
+                    attributes: {
+                        exclude: ['createdAt', 'updatedAt', 'ProductCategory']
+                    }
+                }]
             })
             res.status(200).json(products)
 
@@ -56,7 +79,20 @@ class ProductController {
     static async getProductById(req, res, next) {
         const id = req.params.id
         try {
-            const product = await Product.findByPk(id)
+            const product = await Product.findOne({
+                where: {
+                    id
+                },
+                include: [{
+                    model: Category,
+                    through: {
+                        attributes: []
+                    },
+                    attributes: {
+                        exclude: ['createdAt', 'updatedAt', 'ProductCategory']
+                    }
+                }]
+            })
             res.status(200).json(product)
         } catch (err) {
             next(err)
@@ -65,23 +101,43 @@ class ProductController {
     static async editProduct(req, res, next) {
         try {
             const id = +req.params.id
-            const { name, base_price, category, location } = req.body
+            const { name, description, base_price, category_ids, location } = req.body
+            let categories = category_ids
+            if (typeof categories == 'string') {
+                categories = categories.split(',')
+            }
 
-            var image_url 
+            var image_url
             var image_name
             if (!req.files || Object.keys(req.files).length === 0) {
-               image_url = null
+                image_url = null
             } else {
                 const productExist = await Product.findByPk(id)
-                await app.locals.bucket.file(`products/${productExist.image_name}`).delete()
+                if (productExist.image_name) {
+                    await app.locals.bucket.file(`products/${productExist.image_name}`).delete()
+                }
                 let image = req.files.image;
                 image_name = `PR-${Number(new Date())}-${image.name}`
                 image_url = `https://firebasestorage.googleapis.com/v0/b/market-final-project.appspot.com/o/products%2F${image_name}?alt=media`
                 await app.locals.bucket.file(`products/${image_name}`).createWriteStream().end(req.files.image.data)
             }
+            await ProductCategory.destroy({
+                where:
+                    { product_id: id }
+            })
+
+            Promise.all(
+                categories.map(async el => {
+                    let result = await ProductCategory.create({
+                        category_id: el,
+                        product_id: id
+                    })
+                    return result
+                })
+            )
 
             const updateProduct = await Product.update({
-                name, base_price, category, location, image_url, image_name
+                name, description, base_price, location, image_url, image_name
             }, {
                 where: {
                     id
@@ -99,12 +155,12 @@ class ProductController {
             const productExist = await Product.findAll({
                 include: ['Orders'],
                 where: {
-                   [ Op.and]: [
+                    [Op.and]: [
                         {
                             id
                         },
                         {
-                            '$Orders.status$':'pending'
+                            '$Orders.status$': 'pending'
                         }
                     ]
                 }
@@ -112,13 +168,26 @@ class ProductController {
             if (productExist.length) {
                 next({ name: "redundantOrder" })
             } else {
+                if (productExist.image_name) {
+                    await app.locals.bucket.file(`products/${productExist.image_name}`).delete()
+                }
                 await Product.destroy({
                     where: {
                         id
                     }
                 })
+                await ProductCategory.destroy({
+                    where: {
+                        product_id: id
+                    }
+                })
+                await Order.destroy({
+                    where: {
+                        product_id: id
+                    }
+                })
                 res.status(200).json({
-                    name: "success",
+                    name: "OK",
                     msg: "Product has been deleted"
                 })
             }
